修正圣遗物评分下元素杯分数错误的问题
