+++ apps/profile.js
@@ -1,13 +1,11 @@
-import { Common, App, Data, Cfg } from '../components/index.js'
-import { Character, Player } from '../models/index.js'
-import { getTargetUid, getProfile, profileHelp } from './profile/ProfileCommon.js'
-import { profileArtis, profileArtisList } from './profile/ProfileArtis.js'
-import { renderProfile } from './profile/ProfileDetail.js'
+import { App } from '../components/index.js'
+import { getProfile, profileHelp } from './profile/ProfileCommon.js'
+import { profileArtisList } from './profile/ProfileArtis.js'
+import { profileDetail } from './profile/ProfileDetail.js'
 import { profileStat } from './profile/ProfileStat.js'
 import { profileList } from './profile/ProfileList.js'
 import { uploadCharacterImg, delProfileImg, profileImgList } from './character/ImgUpload.js'
 import { enemyLv } from './profile/ProfileUtils.js'
-import ProfileChange from './profile/ProfileChange.js'
 import { groupRank, resetRank, refreshRank, manageRank } from './profile/ProfileRank.js'
 
 let app = App.init({
@@ -91,119 +89,5 @@ app.reg('profile-img-list', profileImgList, {
   rule: /^#?\s*(.+)(?:面板图列表)\s*$/,
   describe: '【#刻晴面板图列表】 删除指定角色面板图（序号）'
 })
-/**
- app.reg('del-uidflie', delProfile, {
-  rule: /^#?\s*(?:移除|清除|删除)面板数据$/,
-  describe: '【#删除面板数据】 删除面板数据'
-})
- */
 
 export default app
-
-export async function delProfile (e) {
-  let uid = await getTargetUid(e)
-  if (!uid) {
-    return true
-  }
-  if (Data.delfile(`data/UserData/${uid}.json`)) {
-    e.reply(`uid:${uid}缓存面板数据已删除~`)
-  }
-  return true
-}
-
-// 查看当前角色
-export async function profileDetail (e) {
-  let msg = e.original_msg || e.msg
-  if (!msg) {
-    return false
-  }
-  if (!/详细|详情|面板|面版|圣遗物|伤害|换/.test(msg)) {
-    return false
-  }
-  let mode = 'profile'
-  let profileChange = false
-  let changeMsg = msg
-  let pc = ProfileChange.matchMsg(msg)
-  if (pc && pc.char && pc.change) {
-    if (!Cfg.get('profileChange')) {
-      e.reply('面板替换功能已禁用...')
-      return true
-    }
-    e.uid = pc.uid || e.runtime.uid
-    profileChange = ProfileChange.getProfile(e.uid, pc.char, pc.change)
-    if (profileChange && profileChange.char) {
-      msg = `#${profileChange.char?.name}${pc.mode || '面板'}`
-      e._profile = profileChange
-      e._profileMsg = changeMsg
-    }
-  }
-  let uidRet = /[0-9]{9}/.exec(msg)
-  if (uidRet) {
-    e.uid = uidRet[0]
-    msg = msg.replace(uidRet[0], '')
-  }
-
-  let name = msg.replace(/#|老婆|老公/g, '').trim()
-  msg = msg.replace('面版', '面板')
-  let dmgRet = /伤害(\d?)$/.exec(name)
-  let dmgIdx = 0
-  if (/(最强|最高|最高分|最牛|第一)/.test(msg)) {
-    mode = /(分|圣遗物|评分|ACE)/.test(msg) ? 'rank-mark' : 'rank-dmg'
-    name = name.replace(/(最强|最高分|第一|最高|最牛|圣遗物|评分|群)/g, '')
-  }
-  if (/(详情|详细|面板|面版)\s*$/.test(msg) && !/更新|录入|输入/.test(msg)) {
-    mode = 'profile'
-    name = name.replace(/(详情|详细|面板)/, '').trim()
-  } else if (dmgRet) {
-    mode = 'dmg'
-    name = name.replace(/伤害[0-5]?/, '').trim()
-    if (dmgRet[1]) {
-      dmgIdx = dmgRet[1] * 1
-    }
-  } else if (/(详情|详细|面板)更新$/.test(msg) || (/更新/.test(msg) && /(详情|详细|面板)$/.test(msg))) {
-    mode = 'refresh'
-    name = name.replace(/详情|详细|面板|更新/g, '').trim()
-  } else if (/圣遗物/.test(msg)) {
-    mode = 'artis'
-    name = name.replace('圣遗物', '').trim()
-  }
-  if (!Common.cfg('avatarProfile')) {
-    // 面板开关关闭
-    return false
-  }
-  let char = Character.get(name.trim())
-  if (!char) {
-    return false
-  }
-
-  let uid = e.uid || await getTargetUid(e)
-  if (!uid) {
-    return true
-  }
-  e.uid = uid
-  e.avatar = char.id
-
-  if (char.isCustom) {
-    e.reply('自定义角色暂不支持此功能')
-    return true
-  }
-  if (!char.isRelease) {
-    if (!profileChange) {
-      e.reply('角色尚未实装')
-      return true
-    } else if (Cfg.get('notReleasedData') === false) {
-      e.reply('未实装角色面板已禁用...')
-      return true
-    }
-  }
-
-  if (mode === 'profile' || mode === 'dmg') {
-    return renderProfile(e, char, mode, { dmgIdx })
-  } else if (mode === 'refresh') {
-    await getProfile(e)
-    return true
-  } else if (mode === 'artis') {
-    return profileArtis(e)
-  }
-  return true
-}
